+++ apps/web/modules/integrations/webhooks/actions.ts
@@ -1,12 +1,5 @@
 "use server";
 
-import {
-  createWebhook,
-  deleteWebhook,
-  testEndpoint,
-  updateWebhook,
-} from "@/app/(app)/environments/[environmentId]/integrations/webhooks/lib/webhook";
-import { ZWebhookInput } from "@/app/(app)/environments/[environmentId]/integrations/webhooks/types/webhooks";
 import { authenticatedActionClient } from "@/lib/utils/action-client";
 import { checkAuthorizationUpdated } from "@/lib/utils/action-client-middleware";
 import {
@@ -15,6 +8,13 @@ import {
   getProjectIdFromEnvironmentId,
   getProjectIdFromWebhookId,
 } from "@/lib/utils/helper";
+import {
+  createWebhook,
+  deleteWebhook,
+  testEndpoint,
+  updateWebhook,
+} from "@/modules/integrations/webhooks/lib/webhook";
+import { ZWebhookInput } from "@/modules/integrations/webhooks/types/webhooks";
 import { z } from "zod";
 import { ZId } from "@formbricks/types/common";
